refactor(listeners): reorganize command listener packages
- Move command-related listener classes to the `command` package
- Rename `DebugListener` to `DebugCommandListener`
